requirements.txt
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
idna==3.10
lit==18.1.8
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pillow==11.3.0
pycparser==2.23
PyMuPDF==1.26.4
pypdfium2==4.30.0
pytesseract==0.3.13
requests==2.32.5
urllib3==2.5.0
wheel @ file:///usr/local/Cellar/python%403.12/3.12.8/libexec/wheel-0.45.1-py3-none-any.whl#sha256=ff30fbe2536d325272e73cd91fbfbda0f3ea73559a5c26e22e9f5b65429fa0b9
